+++ src/lib/components/table/table-settings/interface/table-settings-dialog-data.interface.ts
@@ -1,0 +1,7 @@
+import { MatColumnDef } from '@angular/material/table';
+import { QueryList } from '@angular/core';
+
+export interface TableSettingsDialogData {
+    columns:Array<MatColumnDef>;
+    selectedColumns:Array<string>;
+}