+++ src/lib/testing/mock-button.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+import { TerraButtonComponent } from '../components/buttons/button/terra-button.component';
+
+@Component({
+    selector: 'terra-button',
+    template: ''
+})
+export class MockButtonComponent extends TerraButtonComponent {}